+++ js_modules/dagit/src/configeditor/types/ConfigEditorPipelineFragment.ts
@@ -12,7 +12,7 @@ export interface ConfigEditorPipelineFragment_environmentType {
 }
 
 export interface ConfigEditorPipelineFragment_types_RegularType {
-  __typename: "RegularType";
+  __typename: "RegularType" | "EnumType";
   name: string;
   isSelector: boolean;
 }
@@ -85,11 +85,12 @@ export enum StepTag {
 }
 
 /**
- * 
+ * This type represents the fields necessary to identify a
+ *         pipeline or pipeline subset.
  */
-export interface PipelineExecutionParams {
-  pipelineName: string;
-  config?: any | null;
+export interface ExecutionSelector {
+  name: string;
+  solidSubset?: string[] | null;
 }
 
 //==============================================================