+++ js_modules/dagit/src/execution/types/PipelineExecutionContainerFragment.ts
@@ -89,7 +89,7 @@ export interface PipelineExecutionContainerFragment_environmentType {
 }
 
 export interface PipelineExecutionContainerFragment_types_RegularType {
-  __typename: "RegularType";
+  __typename: "RegularType" | "EnumType";
   name: string;
   isSelector: boolean;
 }
@@ -163,11 +163,12 @@ export enum StepTag {
 }
 
 /**
- * 
+ * This type represents the fields necessary to identify a
+ *         pipeline or pipeline subset.
  */
-export interface PipelineExecutionParams {
-  pipelineName: string;
-  config?: any | null;
+export interface ExecutionSelector {
+  name: string;
+  solidSubset?: string[] | null;
 }
 
 //==============================================================