+++ js_modules/dagit/src/execution/types/StartPipelineExecution.ts
@@ -110,7 +110,8 @@ export interface StartPipelineExecution {
 }
 
 export interface StartPipelineExecutionVariables {
-  executionParams: PipelineExecutionParams;
+  pipeline: ExecutionSelector;
+  config: any;
 }
 
 /* tslint:disable */
@@ -156,11 +157,12 @@ export enum StepTag {
 }
 
 /**
- * 
+ * This type represents the fields necessary to identify a
+ *         pipeline or pipeline subset.
  */
-export interface PipelineExecutionParams {
-  pipelineName: string;
-  config?: any | null;
+export interface ExecutionSelector {
+  name: string;
+  solidSubset?: string[] | null;
 }
 
 //==============================================================