+++ js_modules/dagit/src/graph/types/SolidNodeFragment.ts
@@ -13,6 +13,7 @@ export interface SolidNodeFragment_definition_metadata {
 }
 
 export interface SolidNodeFragment_definition_configDefinition_type {
+  name: string;
   description: string | null;
 }
 
@@ -130,11 +131,12 @@ export enum StepTag {
 }
 
 /**
- * 
+ * This type represents the fields necessary to identify a
+ *         pipeline or pipeline subset.
  */
-export interface PipelineExecutionParams {
-  pipelineName: string;
-  config?: any | null;
+export interface ExecutionSelector {
+  name: string;
+  solidSubset?: string[] | null;
 }
 
 //==============================================================