+++ python_modules/dagster/dagster/core/execution.py
@@ -633,3 +633,12 @@ def create_typed_environment(pipeline, environment=None):
         raise PipelineConfigEvaluationError(pipeline, result.errors, environment)
 
     return construct_environment_config(result.value)
+
+
+class ExecutionSelector(object):
+    def __init__(self, name, solid_subset=None):
+        self.name = check.str_param(name, 'name')
+        if solid_subset is None:
+            self.solid_subset = None
+        else:
+            self.solid_subset = check.opt_list_param(solid_subset, 'solid_subset', of_type=str)